fix: update register with new group method upgrade: chat backend
