+++ chat-backend/actions/register/app.js
@@ -17,15 +17,17 @@
 
 // ===== ==== ====
 // IMPORTS
-const { GetCommand, UpdateCommand } = require('@aws-sdk/lib-dynamodb') // skipcq: JS-0260
+const { UpdateCommand } = require('@aws-sdk/lib-dynamodb') // skipcq: JS-0260
 
 const { PublishCommand } = require('@aws-sdk/client-sns') // skipcq: JS-0260
 
-const { createVerify } = require('crypto')
+const { verifyIdentity } = require('./src/verify-identity')
 
-const { informGroup, getOtherGroupUsers } = require('./helpers')
+const { informGroup, getOtherGroupUsers } = require('./src/helpers')
 
-const { dynamoDBDocumentClient, snsClient } = require('./aws-clients')
+const { dynamoDBDocumentClient } = require('chat-backend-package/src/clients/aws/dynamo-db-client') // skipcq: JS-0260
+
+const { snsClient } = require('./src/aws-clients')
 
 // ===== ==== ====
 // CONSTANTS
@@ -42,59 +44,13 @@ exports.handler = async (event) => {
 \tRequest Context connectionId: ${event.requestContext.connectionId}
 `)
 
-  const body = JSON.parse(event.body)
-
-  const id = body.id
-  const signature = body.signature
-  const timestamp = body.timestamp
-  let publicKey = body.publicKey // updated later if it already exists
-  if (typeof id === 'undefined' || typeof signature === 'undefined' || typeof timestamp === 'undefined' || typeof publicKey === 'undefined') {
-    throw new Error('id, signature, timestamp, and publicKey must be defined')
+  const validation = await verifyIdentity(event)
+  console.log('\n\n\nIMPORTANT')
+  console.log(validation)
+  if (!validation.isValid) {
+    return validation
   }
-
-  if (Math.abs(Date.now() - timestamp) > 3000) {
-    // prevent repeat attack
-    return {
-      message: 'timestamp is not valid',
-      statusCode: 401
-    }
-  }
-
-  // user
-  const getUserCommand = new GetCommand({
-    TableName: USERS_TABLE_NAME,
-    Key: { id },
-    ProjectionExpression: '#id, #publicKey, #isBanned',
-    ExpressionAttributeNames: {
-      '#id': 'id',
-      '#publicKey': 'publicKey',
-      '#isBanned': 'isBanned'
-    }
-  })
-  const user = await dynamoDBDocumentClient.send(getUserCommand).then((response) => (response.Item))
-  if (typeof user === 'object' && typeof user.publicKey === 'string') {
-    publicKey = user.publicKey
-  }
-
-  // verify is not banned
-  if (typeof user === 'object' && typeof user.isBanned === 'boolean' && user.isBanned) {
-    return {
-      message: 'user is banned',
-      statusCode: 403
-    }
-  }
-
-  // verify signature
-  const verifier = createVerify('rsa-sha256')
-  verifier.update(id + timestamp.toString())
-  const isVerified = verifier.verify(publicKey, Buffer.from(signature), 'base64')
-  console.log('Is the message verified?', isVerified)
-  if (!isVerified) {
-    return {
-      message: 'signature is not valid',
-      statusCode: 403
-    }
-  }
+  const { id, publicKey } = validation
 
   // update user and retrieve old unreadData (if any)
   const updateCommand = new UpdateCommand({
@@ -154,8 +110,9 @@ exports.handler = async (event) => {
   }
 
   // add group
-  registerMessage.groupId = oldUser.groupId ?? oldUser.group // .group for backward compatibility
-  registerMessage.group = registerMessage.groupId // for backward compatibility
+  const groupId = oldUser.groupId ?? oldUser.group // .group for backward compatibility
+  registerMessage.groupId = groupId
+  registerMessage.group = groupId // for backward compatibility
 
   // Retreive other group users
   try {
@@ -167,6 +124,7 @@ exports.handler = async (event) => {
   } catch (e) {
     console.log('Catch: ', e)
   }
+
   console.log('Send Register Message:\n', JSON.stringify(registerMessage))
   // send message
   const publishCommand = new PublishCommand({
