+++ chat-backend/actions/register/package.json
@@ -4,8 +4,11 @@
   "description": "Register user to chat",
   "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "rm -rf dist && esbuild ./app.js --bundle --platform=node --target=node16.14 --outfile=./app.out.js"
   },
   "author": "AdrKacz",
-  "license": "SEE LICENSE IN /LICENSE"
+  "license": "SEE LICENSE IN /LICENSE",
+  "dependencies": {
+    "chat-backend-package": "file:../../chat-backend-package"
+  }
 }
