+++ chat-backend/scripts/pre-deploy/predeploy-package.sh
